Naive jacobi (#26)
* add jacobi

* update cholesky solver

* fix pd

---------

+++ src/simulation/precision.h
@@ -1,1 +1,1 @@
-using solverPrecision = double;
+using solverPrecision = float;
